requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
environs==14.1.1
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
marshmallow==3.26.1
packaging==24.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-telegram-bot==22.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
